Added frontend UI

+++ backend/controllers/goalController.js
@@ -1,12 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 // @desc: Get goals
 // @route GET /api/goals
 // @access: Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -21,7 +23,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -38,6 +41,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found!!')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Checking authorized user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User is not authorized!!')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true,})
     res.status(200).json(updatedGoal)
 })
@@ -51,6 +68,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found!!')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Checking authorized user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User is not authorized!!')
     }
 
     await goal.deleteOne()
